smyrna: use a 'size_t' for 'schemacount'
Squashes a -Wsign-conversion warning. Note that this also required making
`palette.cnt` a `size_t`.

+++ cmd/smyrna/smyrnadefs.h
@@ -107,7 +107,7 @@ typedef struct
     } colorschema;
 
     typedef struct {
-	int schemacount;       /* number of colors */
+	size_t schemacount; // number of colors
 	int smooth;            /* if true, interpolate */
 	colorschema *s;
     } colorschemaset;
